requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
future==0.18.2
herepy==2.1.1
idna==2.9
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
